+++ demo/admin/src/vendors.d.ts
@@ -1,4 +1,5 @@
 /// <reference types="@comet/admin-theme" />
+/// <reference types="vite/client" />
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 declare module "*.svg" {