+++ demo/api/src/pages/entities/page.entity.ts
@@ -1,5 +1,13 @@
 import { BlockDataInterface, RootBlock, RootBlockEntity } from "@comet/blocks-api";
-import { DocumentInterface, PageTreeNodeDocumentEntityScopeService, RootBlockDataScalar, RootBlockType, ScopedEntity } from "@comet/cms-api";
+import {
+    DocumentInterface,
+    EntityInfo,
+    PageTreeNodeDocumentEntityInfoService,
+    PageTreeNodeDocumentEntityScopeService,
+    RootBlockDataScalar,
+    RootBlockType,
+    ScopedEntity,
+} from "@comet/cms-api";
 import { BaseEntity, Entity, OptionalProps, PrimaryKey, Property } from "@mikro-orm/core";
 import { Field, ID, ObjectType } from "@nestjs/graphql";
 import { v4 as uuid } from "uuid";
@@ -7,6 +15,7 @@ import { v4 as uuid } from "uuid";
 import { PageContentBlock } from "../blocks/PageContentBlock";
 import { SeoBlock } from "../blocks/seo.block";
 
+@EntityInfo(PageTreeNodeDocumentEntityInfoService)
 @Entity()
 @ObjectType({
     implements: () => [DocumentInterface],