+++ packages/admin/cms-admin/src/dam/DataGrid/FolderDataGrid.gql.ts
@@ -31,6 +31,7 @@ export const damFileTableFragment = gql`
             ...DamFileThumbnail
         }
         updatedAt
+        createdAt
         importSourceType
     }
     ${damFileThumbnailFragment}
@@ -49,6 +50,7 @@ export const damFolderTableFragment = gql`
         numberOfChildFolders
         isInboxFromOtherScope
         updatedAt
+        createdAt
     }
 `;