+++ packages/site/cms-site/src/blocks/InternalLinkBlock.tsx
@@ -6,15 +6,16 @@ import { PropsWithData } from "./PropsWithData";
 
 interface InternalLinkBlockProps extends PropsWithData<InternalLinkBlockData> {
     children: React.ReactElement;
+    title?: string;
 }
 
-export function InternalLinkBlock({ data: { targetPage, targetPageAnchor }, children }: InternalLinkBlockProps): React.ReactElement {
+export function InternalLinkBlock({ data: { targetPage, targetPageAnchor }, children, title }: InternalLinkBlockProps): React.ReactElement {
     if (!targetPage) {
         return children;
     }
 
     return (
-        <Link href={targetPageAnchor !== undefined ? `${targetPage.path}#${targetPageAnchor}` : targetPage.path} passHref>
+        <Link href={targetPageAnchor !== undefined ? `${targetPage.path}#${targetPageAnchor}` : targetPage.path} passHref title={title}>
             {children}
         </Link>
     );