+++ packages/site/cms-site/src/blocks/factories/OneOfBlock.tsx
@@ -13,7 +13,7 @@ interface Props {
     supportedBlocks: SupportedBlocks;
 }
 
-export const OneOfBlock: React.FC<Props> = ({ data: { block }, supportedBlocks, children }) => {
+export const OneOfBlock: React.FC<Props> = ({ data: { block, ...additionalProps }, supportedBlocks, children }) => {
     if (!block) {
         return null;
     }
@@ -32,5 +32,5 @@ export const OneOfBlock: React.FC<Props> = ({ data: { block }, supportedBlocks, 
         return null;
     }
 
-    return <>{blockFunction({ ...block.props, children })}</>;
+    return <>{blockFunction({ ...block.props, ...additionalProps, children })}</>;
 };